+++ FirstCPPApplication/class.h
@@ -13,10 +13,10 @@ class IClass
         virtual void LevelUpStats(int levels) = 0;
         virtual void LevelUpSkills(int levels) = 0;
 
-        void LevelUpHealth(int change);
-        void LevelUpMana(int change);
-        void LevelUpArmor(int change);
-        void LevelUpDamage(int change);
+        void LevelUpHealth(double change);
+        void LevelUpMana(double change);
+        void LevelUpArmor(double change);
+        void LevelUpDamage(double change);
 };
 
 class FighterClass : public IClass
